# core/api/openapi.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "game/v1/engine.proto",
    "description": "Package game.v1 is the Core service boundary — the central product surface.\nCore serves this contract over both gRPC and REST: the google.api.http\nannotations below drive a grpc-gateway that exposes every RPC as JSON/HTTP\nunder /api/v1, wrapped in the uniform {code, message, trace_id, data}\nenvelope. Core is auth-agnostic — it trusts the caller (a BFF the developer\nbuilds, see examples/) to authenticate and to pass the tenant/merchant\nisolation keys (tenant_id + merchant_id); it only validates and operates on\nthe business object. Errors map to gRPC status codes (and to the envelope at\nthe gateway). Versioned by package; a breaking change becomes game.v2.",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/fulfillment/tasks/{task_id}/callback": {
      "post": {
        "operationId": "FulfillmentService_ReportResult",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int32"
                },
                "message": {
                  "type": "string"
                },
                "trace_id": {
                  "type": "string"
                },
                "data": {
                  "$ref": "#/definitions/v1ReportResultResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/apiErrorEnvelope"
            }
          }
        },
        "parameters": [
          {
            "name": "task_id",
            "description": "Task to finalize (from the path).",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/FulfillmentServiceReportResultBody"
            }
          }
        ],
        "tags": [
          "FulfillmentService"
        ]
      }
    },
...
    "/api/v1/games/{game_id}/eligibility": {
      "get": {
        "operationId": "EngineService_CheckEligibility",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int32"
                },
                "message": {
                  "type": "string"
                },
                "trace_id": {
                  "type": "string"
                },
                "data": {
                  "$ref": "#/definitions/v1CheckEligibilityResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/apiErrorEnvelope"
            }
          }
        },
        "parameters": [
          {
            "name": "game_id",
            "description": "Game to check (from the path).",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "tenant_id",
            "description": "Tenant isolation key (required).",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "merchant_id",
            "description": "Merchant isolation key within the tenant.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "player_id",
            "description": "Player to check.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "EngineService"
        ]
      }
    },
...
    "/api/v1/games/{game_id}/play": {
      "post": {
        "operationId": "EngineService_Play",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int32"
                },
                "message": {
                  "type": "string"
                },
                "trace_id": {
                  "type": "string"
                },
                "data": {
                  "$ref": "#/definitions/v1PlayResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/apiErrorEnvelope"
            }
          }
        },
        "parameters": [
          {
            "name": "game_id",
            "description": "Game being played (from the path).",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EngineServicePlayBody"
            }
          }
        ],
        "tags": [
          "EngineService"
        ]
      }
    },
...
    "/api/v1/games/{game_id}/start": {
      "post": {
        "operationId": "EngineService_StartGame",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int32"
                },
                "message": {
                  "type": "string"
                },
                "trace_id": {
                  "type": "string"
                },
                "data": {
                  "$ref": "#/definitions/v1StartGameResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/apiErrorEnvelope"
            }
          }
        },
        "parameters": [
          {
            "name": "game_id",
            "description": "Game to start (from the path).",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EngineServiceStartGameBody"
            }
          }
        ],
        "tags": [
          "EngineService"
        ]
      }
    },
...
    "/api/v1/leaderboards/{leaderboard_id}/rankings": {
      "get": {
        "operationId": "LeaderboardService_GetRankings",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int32"
                },
                "message": {
                  "type": "string"
                },
                "trace_id": {
                  "type": "string"
                },
                "data": {
                  "$ref": "#/definitions/v1GetRankingsResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/apiErrorEnvelope"
            }
          }
        },
        "parameters": [
          {
            "name": "leaderboard_id",
            "description": "Leaderboard to read (from the path).",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "tenant_id",
            "description": "Tenant isolation key (required).",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "merchant_id",
            "description": "Merchant isolation key within the tenant.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "description": "Max entries to return.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "offset",
            "description": "Offset into the ranking (for paging).",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "LeaderboardService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "EngineServicePlayBody": {
      "type": "object",
      "properties": {
        "tenant_id": {
          "type": "string",
          "description": "Tenant isolation key (required)."
        },
        "merchant_id": {
          "type": "string",
          "description": "Merchant isolation key within the tenant."
        },
        "session_id": {
          "type": "string",
          "description": "Session opened by StartGame."
        },
        "player_id": {
          "type": "string",
          "description": "Player making the play."
        },
        "payload": {
          "type": "object",
          "description": "Shape-specific play input."
        },
        "idempotency_key": {
          "type": "string",
          "description": "Caller-supplied key to make the play idempotent."
        }
      },
      "description": "PlayRequest submits one play within a session."
    },
    "EngineServiceStartGameBody": {
      "type": "object",
      "properties": {
        "tenant_id": {
          "type": "string",
          "description": "Tenant isolation key (required)."
        },
        "merchant_id": {
          "type": "string",
          "description": "Merchant isolation key within the tenant."
        },
        "player_id": {
          "type": "string",
          "description": "Player starting the session."
        }
      },
      "description": "StartGameRequest opens a play session for a player on a game."
    },
    "FulfillmentServiceReportResultBody": {
      "type": "object",
      "properties": {
        "tenant_id": {
          "type": "string",
          "description": "Tenant isolation key (advisory)."
        },
        "merchant_id": {
          "type": "string",
          "description": "Merchant isolation key (advisory)."
        },
        "status": {
          "$ref": "#/definitions/v1TaskStatus",
          "description": "Terminal status: FULFILLED | FAILED."
        },
        "receipt": {
          "type": "string",
          "description": "Raw JSON delivery receipt."
        },
        "error": {
          "type": "string",
          "description": "Error detail, populated when status == FAILED."
        }
      },
      "description": "ReportResult finalizes a task from an external orchestrator (n8n) callback.\nLooked up by the globally-unique task_id; tenant/merchant are advisory."
    },
...
    "v1CheckEligibilityResponse": {
      "type": "object",
      "properties": {
        "can_play": {
          "type": "boolean",
          "description": "Whether the player may play right now."
        },
        "remaining_plays": {
          "type": "integer",
          "format": "int32",
          "description": "Plays remaining under the active caps."
        },
        "reason": {
          "type": "string",
          "description": "Human-readable reason when can_play is false."
        },
        "next_reset_at": {
          "type": "integer",
          "format": "int64",
          "description": "When the play cap next resets (unix seconds; 0 if N/A)."
        }
      }
    },
...
    "v1EntryState": {
      "type": "string",
      "enum": [
        "ENTRY_STATE_UNSPECIFIED",
        "ENTRY_STATE_ACTIVE",
        "ENTRY_STATE_FLAGGED",
        "ENTRY_STATE_DISQUALIFIED"
      ],
      "default": "ENTRY_STATE_UNSPECIFIED",
      "description": "EntryState is a ranked entry's standing."
    },
...
    "v1FulfillmentTask": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "Stable unique identifier (globally unique)."
        },
        "tenant_id": {
          "type": "string",
          "description": "Tenant isolation key."
        },
        "merchant_id": {
          "type": "string",
          "description": "Merchant isolation key within the tenant."
        },
        "reward_id": {
          "type": "string",
          "description": "Reward this task delivers."
        },
        "prize_id": {
          "type": "string",
          "description": "Prize being delivered."
        },
        "player_id": {
          "type": "string",
          "description": "Recipient player."
        },
        "game_id": {
          "type": "string",
          "description": "Game the reward was won in."
        },
        "campaign_id": {
          "type": "string",
          "description": "Campaign the reward belongs to."
        },
        "channel": {
          "type": "string",
          "description": "Delivery channel (e.g. \"webhook\", \"sms\", \"zns\")."
        },
        "channel_config": {
          "type": "string",
          "description": "Raw JSON channel configuration."
        },
        "status": {
          "$ref": "#/definitions/v1TaskStatus",
          "description": "Task lifecycle state."
        },
        "attempts": {
          "type": "integer",
          "format": "int32",
          "description": "Delivery attempts made so far."
        },
        "max_attempts": {
          "type": "integer",
          "format": "int32",
          "description": "Max attempts before the task is dead-lettered."
        },
        "last_error": {
          "type": "string",
          "description": "Last delivery error message."
        },
        "receipt": {
          "type": "string",
          "description": "Raw JSON delivery receipt: voucher code, tracking no., …."
        },
        "next_attempt_at": {
          "type": "integer",
          "format": "int64",
          "description": "Earliest time to retry (unix seconds)."
        },
        "created_at": {
          "type": "integer",
          "format": "int64",
          "description": "Creation time (unix seconds)."
        },
        "updated_at": {
          "type": "integer",
          "format": "int64",
          "description": "Last update time (unix seconds)."
        }
      },
      "description": "FulfillmentTask is one row of the transactional outbox: a unit of delivery\nwork committed atomically with the reward it delivers."
    },
...
    "v1GetRankingsResponse": {
      "type": "object",
      "properties": {
        "entries": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1RankedEntry"
          },
          "description": "Page of ranked entries."
        },
        "total": {
          "type": "string",
          "format": "int64",
          "description": "Total ranked entries."
        },
        "window_key": {
          "type": "string",
          "description": "Key identifying the active time window."
        }
      }
    },
...
    "v1PlayResponse": {
      "type": "object",
      "properties": {
        "play_id": {
          "type": "string",
          "description": "Durable id of this play."
        },
        "rewards": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Reward"
          },
          "description": "Rewards won on this play (may be empty)."
        },
        "metadata": {
          "type": "object",
          "description": "Shape-specific result metadata."
        }
      }
    },
...
    "v1RankedEntry": {
      "type": "object",
      "properties": {
        "player_id": {
          "type": "string",
          "description": "Player the entry belongs to."
        },
        "score": {
          "type": "string",
          "format": "int64",
          "description": "Ranking score."
        },
        "plays": {
          "type": "string",
          "format": "int64",
          "description": "Plays contributing to the score."
        },
        "rank": {
          "type": "string",
          "format": "int64",
          "description": "1-based rank."
        },
        "state": {
          "$ref": "#/definitions/v1EntryState",
          "description": "Entry standing (active/flagged/disqualified)."
        }
      },
      "description": "RankedEntry is one player's standing (player list / rankings)."
    },
...
    "v1ReportResultResponse": {
      "type": "object",
      "properties": {
        "task": {
          "$ref": "#/definitions/v1FulfillmentTask",
          "description": "The finalized task."
        }
      }
    },
...
    "v1Reward": {
      "type": "object",
      "properties": {
        "prize_id": {
          "type": "string",
          "description": "Prize that was awarded."
        },
        "name": {
          "type": "string",
          "description": "Prize name for display."
        },
        "image": {
          "type": "string",
          "description": "Prize image URL for display."
        },
        "type": {
          "type": "string",
          "description": "Prize type."
        },
        "quantity": {
          "type": "string",
          "format": "int64",
          "description": "Number of units awarded."
        },
        "value": {
          "type": "string",
          "format": "int64",
          "description": "Nominal value for display."
        },
        "reward_id": {
          "type": "string",
          "description": "Durable reward record id (used to claim)."
        },
        "code": {
          "type": "string",
          "description": "Assigned voucher code, if any."
        }
      },
      "description": "Reward is a single awarded item in a play result."
    },
...
    "v1StartGameResponse": {
      "type": "object",
      "properties": {
        "session_id": {
          "type": "string",
          "description": "Opaque session id to pass to Play."
        },
        "seed_data": {
          "type": "string",
          "description": "Raw JSON seed data for the client (may be empty)."
        },
        "expires_at": {
          "type": "integer",
          "format": "int64",
          "description": "Session expiry (unix seconds)."
        }
      }
    },
    "v1TaskStatus": {
      "type": "string",
      "enum": [
        "TASK_STATUS_UNSPECIFIED",
        "TASK_STATUS_PENDING",
        "TASK_STATUS_PROCESSING",
        "TASK_STATUS_FULFILLED",
        "TASK_STATUS_FAILED",
        "TASK_STATUS_DEAD"
      ],
      "default": "TASK_STATUS_UNSPECIFIED",
      "description": "TaskStatus is the state of a fulfillment outbox task. ReportResult callbacks\nfinalize a task to FULFILLED or FAILED."
    },
...
    "apiFieldViolation": {
      "type": "object",
      "properties": {
        "field": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "apiErrorInfo": {
      "type": "object",
      "description": "data.error payload (mirrors google.rpc.ErrorInfo).",
      "properties": {
        "status": {
          "type": "string",
          "description": "canonical gRPC status name"
        },
        "reason": {
          "type": "string",
          "description": "stable machine reason (see docs/ERRORS.md)"
        },
        "domain": {
          "type": "string"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "field_violations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiFieldViolation"
          }
        }
      }
    },
    "apiErrorData": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/apiErrorInfo"
        }
      }
    },
    "apiErrorEnvelope": {
      "type": "object",
      "description": "Uniform error envelope.",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "trace_id": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/apiErrorData"
        }
      }
    }
  }
}